Some cleanup from PR feedback.
We added a proper fallback to 'convert_legavy_static_url_with_course_id'
if we can't find the asset from the AssetManager.  This gives us better
  confidence in the exception case to 'do the right thing'.

We also added more tests surrounding thumbnails and static asset URLs.
There was also a lot of cleanup of the test cases to not visually be so
cryptic, and they should be easier to understand now.