More consolidation of helper code.
Removed some legacy stuff and consolidated further into
get_canonicalized_asset_path.  This makes it a little simpler to reason
about, I think, since we're not falling back to a method which did
almost the exact same thing.  If we can't find an asset and get its
metadata, we treat it as locked, that's it.

Updated the tests to not only explain what each section does, but to
also add more cases -- subdirectories, query params -- and make them
much more solid.